fix(client): Restore client mod and get it compiling
Whilst it compiles, the implementation is not finished. Current issues
include:

1. Verbose and duplicated serialisation logic in `Session`.
2. Still requires duplicate serialisation to succeed (once to generate
   signature of serialised message, once again to serialise message
   before sending).
3. Impossible to generate a signature over the whole message, since
   message ID is generated by send function and stored in message.

Issues 2 & 3 are pre-existing, but are technically avoided since we
don't generate real signatures (we just sign `b"TODO"`). Issue 1 could
easily be fixed with a helper `fn`, but this would likely need to be
broken up again when it comes time to actually implement signatures, so
it's been left as-is for now.